+++ continuous_integration/run_tests.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 
-set -e
+set -xe
 
 echo "Activating test environment:"
 if [[ "$PYTHON_VERSION" == pypy3* ]]; then
     source pypy3/bin/activate
+elif [[ "$PYTHON_VERSION" == free-threaded* ]]; then
+    source testenv/bin/activate
+    # This is needed because for now some C extensions have not declared their
+    # thread-safety with free-threaded Python, for example numpy and coverage.tracer
+    export PYTHON_GIL=0
 else
     source activate testenv
 fi